#131 [FEAT] 상수 & 유틸 파일로 분리
